% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display_orange_all}
\alias{display_orange_all}
\title{Display all the available color palettes}
\usage{
display_orange_all()
}
\description{
Display all the available color palettes simultanueously in a graphics
window.
}
\examples{
display_orange_all()
}
\seealso{
\code{\link{get_orange_cols}}, \code{\link{orange_palettes}},
  \code{\link{display_orange_palette}}
}
