% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display_orange_palette}
\alias{display_orange_palette}
\title{Display a color palette}
\usage{
display_orange_palette(palette, n = NULL)
}
\arguments{
\item{palette}{Character name of a palette. One of "principale",
"secondaire", "gris", "bleus", "verts", "roses", "violets" or "jaunes".}

\item{n}{Numeric indicating the number of different colors in the palette. If
NULL (the default) all the colors of the palette are displayed.}
}
\description{
Display a color palette in a graphics window.
}
\examples{
display_orange_palette("principale")
display_orange_palette("bleus", n = 10)
}
\seealso{
\code{\link{get_orange_cols}}, \code{\link{orange_palettes}},
  \code{\link{display_orange_all}}
}
