% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{list_orange_palettes}
\alias{list_orange_palettes}
\title{Orange Color Palettes}
\format{An object of class \code{list} of length 8.}
\usage{
list_orange_palettes
}
\description{
A collection of color palettes that respect the corporate identity and style
guide of Orange™. See
\url{https://www.ateliers-orange.fr/index.php?page=marque} for more details.
}
\keyword{datasets}
