% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{orange_colors}
\alias{orange_colors}
\title{Extract Orange's colors}
\usage{
orange_colors(col = c("gris foncé", "gris moyen", "gris clair", "orange",
  "bleu foncé", "bleu", "bleu clair", "vert foncé", "vert", "vert clair",
  "rose foncé", "rose", "rose clair", "violet foncé", "violet",
  "violet clair", "jaune foncé", "jaune", "jaune clair"))
}
\arguments{
\item{col}{Character names of colors.}
}
\description{
Extract Orange's colors as hexadecimal codes.
}
\examples{
orange_colors()
orange_colors("orange")
orange_colors(c("bleu", "vert"))

# you can use the colors manually in plots
ggplot(mtcars, aes(hp, mpg)) +
  geom_point(color = orange_colors("vert"), size = 4, alpha = .8) +
  theme_minimal()

ggplot(mpg, aes(x = cty, y = displ, color = fl)) +
  geom_point(size = 4, alpha = 0.8) +
  scale_color_manual(
    values = unname(
      orange_colors(c("orange", "bleu", "jaune", "violet", "vert"))
    )
  ) +
  theme_minimal()
}
\seealso{
\code{\link{orange_palettes}}, \code{\link{scale_color_orange}},
  \code{\link{scale_fill_orange}}
}
