% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{orange_palettes}
\alias{orange_palettes}
\title{Interpolate a color palette}
\usage{
orange_palettes(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette. One of "main",
"secondary", "grey", "blue", "green", "pink", "purple" or "yellow".}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to colorRampPalette().}
}
\description{
Interpolate a color palette.
}
\examples{
orange_palettes("main")(2)
orange_palettes("secondary", reverse = TRUE)(6)
orange_palettes("blue")(10)
}
\seealso{
\code{\link{orange_colors}}, \code{\link{scale_color_orange}},
  \code{\link{scale_fill_orange}}
}
