% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_orange}
\alias{scale_color_orange}
\title{Color scale constructor}
\usage{
scale_color_orange(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of a palette. One of "main",
"secondary", "grey", "blue", "green", "pink", "purple" or "yellow".}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE.}
}
\description{
This function specifies a palette, whether the palette is being applied based
on a discrete or numeric variable, whether to reverse the palette colors, and
additional arguments to pass to the relevant ggplot2 function (which differs
for discrete or numeric mapping).
}
\examples{
library("ggplot2")

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_orange(palette = "secondary", discrete = TRUE) +
  theme_minimal()

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Petal.Length)) +
  geom_point(size = 4) +
  scale_color_orange(palette = "blue", discrete = FALSE) +
  theme_minimal()
}
\seealso{
\code{\link{orange_colors}}, \code{\link{orange_palettes}},
  \code{\link{scale_fill_orange}}
}
