% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_orange}
\alias{scale_fill_orange}
\title{Fill scale constructor}
\usage{
scale_fill_orange(palette = "principale", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette. One of "principale",
"secondaire", "gris", "bleus", "verts", "roses", "violets" or "jaunes".}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE.}
}
\description{
This function specifies a palette, whether the palette is being applied based
on a discrete or numeric variable, whether to reverse the palette colors, and
additional arguments to pass to the relevant ggplot2 function (which differs
for discrete or numeric mapping).
}
\examples{
ggplot(mpg, aes(x = fl, y = displ, fill = fl)) +
  geom_boxplot() +
  scale_fill_orange(palette = "secondaire", guide = "none") +
  theme_minimal()

ggplot(mpg, aes(x = manufacturer, fill = manufacturer)) +
  geom_bar() +
  scale_fill_orange(palette = "principale", guide = "none") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
}
\seealso{
\code{\link{get_orange_cols}}, \code{\link{orange_palettes}},
  \code{\link{scale_color_orange}}
}
