% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{theme_elegant}
\alias{theme_elegant}
\title{A simple and elegant black and white theme}
\usage{
theme_elegant(base_family = "", base_size = 14, title_size = 18,
  subtitle_size = 14, caption_size = 11, facet_text_size = 12,
  axis_size = 12, axis_title_position = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30))
}
\arguments{
\item{base_family}{Base font family.}

\item{base_size}{Base font size.}

\item{title_size}{Title font size.}

\item{subtitle_size}{Subtitle font size.}

\item{caption_size}{Caption font size.}

\item{facet_text_size}{Facet font size.}

\item{axis_size}{Axis font size.}

\item{axis_title_position}{Axis title position. One of `[blmcrt]`.}

\item{plot_margin}{Margins of the plot.}
}
\description{
A simple and elegant black and white theme.
}
\examples{
library("ggplot2")

ggplot(data = iris,
       aes(x = Sepal.Width, y = Sepal.Length, color = Species)) +
  geom_point(size = 3, show.legend = FALSE) +
  geom_smooth(se = FALSE, show.legend = FALSE) +
  scale_color_orange(palette = "secondary") +
  facet_wrap(~ Species) +
  labs(title = "The 'infamous' iris dataset",
  subtitle = "Sepal.Length in function of Sepal.Width by Species") +
  theme_elegant()
}
