% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{orange_colors}
\alias{orange_colors}
\title{Extract Orange's colors}
\usage{
orange_colors(col = c("dark grey", "grey", "light grey", "orange",
  "dark blue", "blue", "light blue", "dark green", "green", "light green",
  "dark pink", "pink", "light pink", "dark purple", "purple", "light purple",
  "dark yellow", "yellow", "light yellow"))
}
\arguments{
\item{col}{Character names of colors.}
}
\description{
Extract Orange's colors as hexadecimal codes.
}
\examples{
orange_colors()
orange_colors("orange")
orange_colors(c("blue", "green"))

# you can use the colors manually in plots
library("ggplot2")

ggplot(mtcars, aes(hp, mpg)) +
  geom_point(color = orange_colors("green"), size = 4, alpha = .8) +
  theme_minimal()

ggplot(mpg, aes(x = cty, y = displ, color = fl)) +
  geom_point(size = 4, alpha = 0.8) +
  scale_color_manual(
    values = unname(
      orange_colors(c("orange", "blue", "yellow", "purple", "green"))
    )
  ) +
  theme_minimal()
}
\seealso{
\code{\link{orange_palettes}}, \code{\link{scale_color_orange}},
  \code{\link{scale_fill_orange}}
}
