% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{orange_palettes}
\alias{orange_palettes}
\title{Interpolate a color palette}
\usage{
orange_palettes(palette = "principale", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette. One of "principale",
"secondaire", "gris", "bleus", "verts", "roses", "violets" or "jaunes".}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to colorRampPalette().}
}
\description{
Interpolate a color palette.
}
\examples{
orange_palettes("principale")(2)
orange_palettes("secondaire", reverse = TRUE)(6)
orange_palettes("bleus")(10)
}
\seealso{
\code{\link{get_orange_cols}}, \code{\link{scale_color_orange}},
  \code{\link{scale_fill_orange}}
}
